refactor: extract chat presentation from AgentWorkspace
